renaming the config-server module adding integration test and adding integration test module and defining deployment folder
