+++ config-server/src/main/resources/application.yaml
@@ -1,0 +1,2 @@
+server:
+    port:8071
