+++ config-server/src/test/java/com/ejazz/config_service/ConfigServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package com.ejazz.config_service;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ConfigServiceApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
