--- config-service/src/main/java/com/ejazz/config_service/ConfigServiceApplication.java
@@ -1,14 +1,0 @@
-package com.ejazz.config_service;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-@enableConfigServer
-public class ConfigServiceApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ConfigServiceApplication.class, args);
-	}
-
-}
