--- config-service/src/main/resources/application.yaml
@@ -1,2 +1,0 @@
-server:
-    port:8071
