--- config-service/src/test/java/com/ejazz/config_service/ConfigServiceApplicationTests.java
@@ -1,13 +1,0 @@
-package com.ejazz.config_service;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ConfigServiceApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
