+++ integration-tests/src/main/java/com/ejazz/integration_tests/IntegrationTestsApplication.java
@@ -1,0 +1,13 @@
+package com.ejazz.integration_tests;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class IntegrationTestsApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(IntegrationTestsApplication.class, args);
+	}
+
+}
