+++ integration-tests/src/test/java/com/ejazz/integration_tests/IntegrationTestsApplicationTests.java
@@ -1,0 +1,15 @@
+package com.ejazz.integration_tests;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.context.annotation.Import;
+
+@Import(TestcontainersConfiguration.class)
+@SpringBootTest
+class IntegrationTestsApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
