+++ integration-tests/src/test/java/com/ejazz/integration_tests/TestIntegrationTestsApplication.java
@@ -1,0 +1,11 @@
+package com.ejazz.integration_tests;
+
+import org.springframework.boot.SpringApplication;
+
+public class TestIntegrationTestsApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.from(IntegrationTestsApplication::main).with(TestcontainersConfiguration.class).run(args);
+	}
+
+}
