+++ integration-tests/src/test/java/com/ejazz/integration_tests/TestcontainersConfiguration.java
@@ -1,0 +1,8 @@
+package com.ejazz.integration_tests;
+
+import org.springframework.boot.test.context.TestConfiguration;
+
+@TestConfiguration(proxyBeanMethods = false)
+class TestcontainersConfiguration {
+
+}
