+++ pom.xml
@@ -11,6 +11,7 @@
         <module>common</module>
         <module>account-service</module>
         <module>post-service</module>
+        <module>config-server</module>
     </modules>
 
     <dependencyManagement>
